Calculator and Fraction changed

+++ Blatt 4 A1/src/fraction/Fraction.java
@@ -169,7 +169,7 @@ public class Fraction extends Number {
 
    @Override
    public int intValue() {
-      return (int)numerator / denominator;
+      return numerator / denominator;
    }
 
    @Override
